requirements.txt
absl-py==0.6.1
aiohttp==3.5.4
asn1crypto==0.24.0
astor==0.7.1
astroid==2.2.5
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
Automat==0.7.0
autopep8==1.4.3
backcall==0.1.0
bcolz==1.2.1
bcrypt==3.1.6
beautifulsoup4==4.7.1
bleach==3.0.2
blinker==1.4
bokeh==1.0.1
Bottleneck==1.2.1
cachetools==3.1.0
certifi==2018.10.15
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==0.8.0
constantly==15.1.0
coverage==4.5.3
cryptography==2.6.1
cssselect==1.0.3
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
dask==1.1.3
dataclasses==0.6
decorator==4.4.0
defusedxml==0.5.0
dill==0.2.8.2
distributed==1.26.0
Django==2.2
entrypoints==0.2.3
fastai==1.0.38
fastprogress==0.1.18
feather-format==0.4.0
flake8==3.6.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Session==0.3.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gast==0.2.0
google==2.0.1
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
graphviz==0.10.1
grpcio==1.16.0
h5py==2.8.0
HeapDict==1.0.0
html5lib==1.0.1
httplib2==0.12.1
hyperlink==19.0.0
idna==2.8
idna-ssl==1.1.0
incremental==17.5.0
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.17
isoweek==1.3.3
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
locket==0.2.0
lorem==0.1.1
lxml==4.3.3
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.0.0
mpl-finance==0.10.0
msgpack==0.5.6
msgpack-numpy==0.4.3.2
multidict==4.5.2
murmurhash==1.0.1
mysql==0.0.2
mysqlclient==1.4.2.post1
nbconvert==5.4.0
nbformat==4.4.0
nltk==3.3
notebook==5.7.0
numexpr==2.6.8
numpy==1.15.4
nvidia-ml-py3==7.352.0
opencv-python==3.4.4.19
packaging==18.0
pandarallel==0.1.1
pandas==0.23.4
pandas-summary==0.0.5
pandocfilters==1.4.2
parsel==1.5.1
parso==0.4.0
partd==0.3.9
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.3.0
pkg-resources==0.0.0
plac==0.9.6
pluggy==0.9.0
preshed==2.0.1
prometheus-client==0.4.2
prompt-toolkit==2.0.9
protobuf==3.6.1
psutil==5.5.1
ptyprocess==0.6.0
py==1.8.0
pyarrow==0.12.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.4.0
pycparser==2.19
PyDispatcher==2.0.5
pyflakes==2.0.0
Pygments==2.3.1
PyHamcrest==1.9.0
pylint==2.3.1
pyOpenSSL==19.0.0
pyparsing==2.3.0
pytest==4.4.1
pytest-cov==2.6.1
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
queuelib==1.5.0
regex==2018.1.10
requests==2.20.1
rsa==4.0
scikit-learn==0.20.0
scipy==1.1.0
Scrapy==1.6.0
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.12.0
sklearn==0.0
sklearn-pandas==1.8.0
sortedcontainers==2.1.0
soupsieve==1.9.1
spacy==2.0.16
SQLAlchemy==1.3.3
sqlparse==0.3.0
svgwrite==1.2.1
tailer==0.4.1
tblib==1.3.2
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
thinc==6.12.0
toolz==0.9.0
torch==0.4.1
torchvision==0.2.1
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
Twisted==19.2.0
typed-ast==1.3.1
typing==3.6.6
typing-extensions==3.7.2
ujson==1.35
uritemplate==3.0.0
urllib3==1.24.2
w3lib==1.20.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
widgetsnbextension==3.4.2
wrapt==1.10.11
WTForms==2.2.1
yarl==1.3.0
zict==0.1.3
zope.interface==4.6.0
PyPDF2==1.26.0
pdf2image
Pillow
pyvirtualdisplay
